+++ CHANGELOG.md
@@ -12,6 +12,13 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 - Exonum nodes now gracefully terminate on receiving SIGINT or SIGTERM signals
   (on Unix platforms), or a `ctrl + c` break (on Windows). (#1834)
 
+### Bug Fixes
+
+#### exonum-cli
+
+- Fixed bug in `run-dev` command. Previously, using the existing configuration
+  files in `blockchain-path` led to an error. (#1827)
+
 ## 1.0.0-rc.3 - 2020-03-25
 
 ### Breaking changes